+++ pyatv/protocols/companion/protocol.py
@@ -9,13 +9,13 @@ from pyatv import exceptions
 from pyatv.auth.hap_pairing import parse_credentials
 from pyatv.auth.hap_srp import SRPAuthHandler
 from pyatv.interface import BaseService
-from pyatv.protocols.companion import opack
 from pyatv.protocols.companion.auth import CompanionPairVerifyProcedure
 from pyatv.protocols.companion.connection import (
     CompanionConnection,
     CompanionConnectionListener,
     FrameType,
 )
+from pyatv.support import opack
 from pyatv.support.collections import SharedData
 from pyatv.support.state_producer import StateProducer
